西元2025年03月12日 (週三) 13時15分03秒 CST

+++ nextcloud_server/server/apps/theming/lib/Service/BackgroundService.php
@@ -20,7 +20,6 @@ use OCP\Files\SimpleFS\ISimpleFile;
 use OCP\Files\SimpleFS\ISimpleFolder;
 use OCP\IAppConfig;
 use OCP\IConfig;
-use OCP\Image;
 use OCP\Lock\LockedException;
 use OCP\PreConditionNotMetException;
 use RuntimeException;
@@ -209,7 +208,10 @@ class BackgroundService {
 	}
 
 	public function setDefaultBackground(?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
+		$userId = $userId ?? $this->userId;
+		if ($userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
 
 		$this->config->deleteUserValue($userId, Application::APP_ID, 'background_image');
 		$this->config->deleteUserValue($userId, Application::APP_ID, 'background_color');
@@ -224,9 +226,11 @@ class BackgroundService {
 	 * @throws PreConditionNotMetException
 	 * @throws NoUserException
 	 */
-	public function setFileBackground(string $path, ?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
-		$userFolder = $this->rootFolder->getUserFolder($userId);
+	public function setFileBackground($path): void {
+		if ($this->userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
+		$userFolder = $this->rootFolder->getUserFolder($this->userId);
 
 		/** @var File $file */
 		$file = $userFolder->get($path);
@@ -240,9 +244,12 @@ class BackgroundService {
 	}
 
 	public function recalculateMeanColor(?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
+		$userId = $userId ?? $this->userId;
+		if ($userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
 
-		$image = new Image();
+		$image = new \OCP\Image();
 		$handle = $this->getAppDataFolder($userId)->getFile('background.jpg')->read();
 		if ($handle === false || $image->loadFromFileHandle($handle) === false) {
 			throw new InvalidArgumentException('Invalid image file');
@@ -263,8 +270,10 @@ class BackgroundService {
 	 * @throws InvalidArgumentException If the specified filename does not match any shipped background
 	 */
 	public function setShippedBackground(string $filename, ?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
-
+		$userId = $userId ?? $this->userId;
+		if ($userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
 		if (!array_key_exists($filename, self::SHIPPED_BACKGROUNDS)) {
 			throw new InvalidArgumentException('The given file name is invalid');
 		}
@@ -278,8 +287,10 @@ class BackgroundService {
 	 * @param string|null $userId The user to set the color - default to current logged-in user
 	 */
 	public function setColorBackground(string $color, ?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
-
+		$userId = $userId ?? $this->userId;
+		if ($userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
 		if (!preg_match('/^#([0-9a-f]{3}|[0-9a-f]{6})$/i', $color)) {
 			throw new InvalidArgumentException('The given color is invalid');
 		}
@@ -287,18 +298,19 @@ class BackgroundService {
 		$this->config->setUserValue($userId, Application::APP_ID, 'background_image', self::BACKGROUND_COLOR);
 	}
 
-	public function deleteBackgroundImage(?string $userId = null): void {
-		$userId = $userId ?? $this->getUserId();
-		$this->config->setUserValue($userId, Application::APP_ID, 'background_image', self::BACKGROUND_COLOR);
+	public function deleteBackgroundImage(): void {
+		if ($this->userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
+		$this->config->setUserValue($this->userId, Application::APP_ID, 'background_image', self::BACKGROUND_COLOR);
 	}
 
-	public function getBackground(?string $userId = null): ?ISimpleFile {
-		$userId = $userId ?? $this->getUserId();
-		$background = $this->config->getUserValue($userId, Application::APP_ID, 'background_image', self::BACKGROUND_DEFAULT);
+	public function getBackground(): ?ISimpleFile {
+		$background = $this->config->getUserValue($this->userId, Application::APP_ID, 'background_image', self::BACKGROUND_DEFAULT);
 		if ($background === self::BACKGROUND_CUSTOM) {
 			try {
 				return $this->getAppDataFolder()->getFile('background.jpg');
-			} catch (NotFoundException|NotPermittedException $e) {
+			} catch (NotFoundException | NotPermittedException $e) {
 				return null;
 			}
 		}
@@ -311,8 +323,8 @@ class BackgroundService {
 	 * @param resource|string $path
 	 * @return string|null The fallback background color - if any
 	 */
-	public function setGlobalBackground($path): ?string {
-		$image = new Image();
+	public function setGlobalBackground($path): string|null {
+		$image = new \OCP\Image();
 		$handle = is_resource($path) ? $path : fopen($path, 'rb');
 
 		if ($handle && $image->loadFromFileHandle($handle) !== false) {
@@ -329,7 +341,7 @@ class BackgroundService {
 	 * Calculate mean color of an given image
 	 * It only takes the upper part into account so that a matching text color can be derived for the app menu
 	 */
-	private function calculateMeanColor(Image $image): false|string {
+	private function calculateMeanColor(\OCP\Image $image): false|string {
 		/**
 		 * Small helper to ensure one channel is returned as 8byte hex
 		 */
@@ -337,13 +349,13 @@ class BackgroundService {
 			$hex = dechex($channel);
 			return match (strlen($hex)) {
 				0 => '00',
-				1 => '0' . $hex,
+				1 => '0'.$hex,
 				2 => $hex,
 				default => 'ff',
 			};
 		}
 
-		$tempImage = new Image();
+		$tempImage = new \OCP\Image();
 
 		// Crop to only analyze top bar
 		$resource = $image->cropNew(0, 0, $image->width(), min(max(50, (int)($image->height() * 0.125)), $image->height()));
@@ -388,27 +400,20 @@ class BackgroundService {
 	 * @throws NotPermittedException
 	 */
 	private function getAppDataFolder(?string $userId = null): ISimpleFolder {
-		$userId = $userId ?? $this->getUserId();
+		$userId = $userId ?? $this->userId;
+		if ($userId === null) {
+			throw new RuntimeException('No currently logged-in user');
+		}
 
 		try {
 			$rootFolder = $this->appData->getFolder('users');
-		} catch (NotFoundException) {
+		} catch (NotFoundException $e) {
 			$rootFolder = $this->appData->newFolder('users');
 		}
 		try {
 			return $rootFolder->getFolder($userId);
-		} catch (NotFoundException) {
+		} catch (NotFoundException $e) {
 			return $rootFolder->newFolder($userId);
 		}
-	}
-
-	/**
-	 * @throws RuntimeException Thrown if a method that needs a user is called without any logged-in user
-	 */
-	private function getUserId(): string {
-		if ($this->userId === null) {
-			throw new RuntimeException('No currently logged-in user');
-		}
-		return $this->userId;
 	}
 }
